Docker tweaking
Updated docker compose file to v2.  After some testing, hostfile does not work as expected in docker which is causing issues with getting nginx to read from the app container.  I found an existing project that fixes this using docker-gen.  Might be something t oconsider down the road.  Going to merge app and nginx containers for now to get this running.